Remove InMemoryVaultStore
