--- Vault/Sources/VaultFeed/Storage/InMemoryVaultStore.swift
@@ -1,93 +1,0 @@
-import Foundation
-import RegexBuilder
-
-public final actor InMemoryVaultStore {
-    private var codes: [VaultItem]
-
-    public init(codes: [VaultItem] = []) {
-        self.codes = codes
-    }
-}
-
-extension InMemoryVaultStore: VaultStoreReader {
-    public func retrieve() async throws -> VaultRetrievalResult<VaultItem> {
-        .init(items: codes, errors: [])
-    }
-
-    public func retrieve(matching query: String) async throws -> VaultRetrievalResult<VaultItem> {
-        let pattern = Regex {
-            ZeroOrMore { .any }
-            query
-            ZeroOrMore { .any }
-        }.ignoresCase()
-        let codes = codes.filter { $0.matches(pattern: pattern) }
-        return .init(items: codes, errors: [])
-    }
-}
-
-extension VaultItem {
-    fileprivate func matches(pattern: Regex<Substring>) -> Bool {
-        let fields: [String?] = [
-            metadata.userDescription,
-            item.otpCode?.data.accountName,
-            item.otpCode?.data.issuer,
-            item.secureNote?.title,
-            item.secureNote?.contents,
-        ]
-        return fields.contains(where: { $0?.contains(pattern) == true })
-    }
-}
-
-extension InMemoryVaultStore: VaultStoreWriter {
-    /// Thrown if a code cannot be found for a given operation.
-    struct CodeNotFound: Error {}
-
-    @discardableResult
-    public func insert(item: VaultItem.Write) async throws -> UUID {
-        let currentDate = Date()
-        let metadata = VaultItem.Metadata(
-            id: UUID(),
-            created: currentDate,
-            updated: currentDate,
-            userDescription: item.userDescription,
-            tags: item.tags,
-            visibility: item.visibility,
-            searchableLevel: item.searchableLevel,
-            searchPassphrase: item.searchPassphase,
-            color: item.color
-        )
-        let code = VaultItem(
-            metadata: metadata,
-            item: item.item
-        )
-        codes.append(code)
-        return code.id
-    }
-
-    public func update(id: UUID, item: VaultItem.Write) async throws {
-        guard let index = codes.firstIndex(where: { $0.id == id }) else {
-            throw CodeNotFound()
-        }
-        let existingCode = codes[index]
-        let metadata = VaultItem.Metadata(
-            id: id,
-            created: existingCode.metadata.created,
-            updated: Date(),
-            userDescription: item.userDescription,
-            tags: item.tags,
-            visibility: item.visibility,
-            searchableLevel: item.searchableLevel,
-            searchPassphrase: item.searchPassphase,
-            color: item.color
-        )
-        let newCode = VaultItem(
-            metadata: metadata,
-            item: item.item
-        )
-        codes[index] = newCode
-    }
-
-    public func delete(id: UUID) async throws {
-        codes.removeAll(where: { $0.id == id })
-    }
-}
